Improve developer workflow with systemd flag
As the working directory was moved into /run/faasd we were
unable to run faasd from our current working directory as
various paths were hard coded to this new directory and the
faasd command assumed it was being run from there.

When developing its easier to run from the go path as that
is where we build the binary. This commit allows passing a
flag to the install command to install only the required
files into the /run/faasd directory, and adds a flag to the
up command that allows us to fake running from that directory

This has been tested by running install without the new flag,
the services get created and installed. These were removed,
disabled and the /run/faasd directory was deleted. Then
install was run with the new --install-services flag set to
false, which should only create and install the files into the
/run/faasd directory.

The up command was tested with the new --work-dir flag pointing
at the /run/faasd directory and it worked when $PWD was not
/run/faasd. (faas-containerd run manually in another terminal)

Signed-off-by: Alistair Hey <alistair@heyal.co.uk>